backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose
passlib[bcrypt]
python-dotenv
python-multipart